Refactor: page와 component간의 package 구조 수정
