+++ src/components/example1/Example1.tsx
@@ -1,0 +1,9 @@
+const Example1 = () => {
+  return (
+    <>
+      <h1>main 페이지에 중첩된 example page</h1>
+    </>
+  );
+};
+
+export default Example1;
