+++ src/components/example2/Example2.tsx
@@ -1,0 +1,9 @@
+const Example2 = () => {
+  return (
+    <>
+      <h1>Example2 page</h1>
+    </>
+  );
+};
+
+export default Example2;
