--- src/pages/Example1.tsx
@@ -1,10 +1,0 @@
-import Example1 from './example1/Example1';
-
-const Example1Page = () => {
-
-  return (
-    <Example1 />
-  );
-}
-
-export default Example1Page;
