--- src/pages/Example2.tsx
@@ -1,9 +1,0 @@
-import Example2 from './example2/Example2';
-
-const Example2Page = () => {
-  return (
-    <Example2 />
-  );
-};
-
-export default Example2Page;
